VAS_Projekt/requirements.txt
spade>=3.3
PyYAML>=6.0
